requirements.txt
pyTelegramBotAPI
requests
openai
python-dotenv
python-dateutil
pytz